refactor: move source and dest network containers out
